+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 * Added a composer.json for PHP packages ([view](https://github.com/videojs/video.js/pull/1241))
 * Exposed the vertical option for slider controls ([view](https://github.com/videojs/video.js/pull/1303))
 * Fixed an error when disposing a tech using manual timeupdates ([view](https://github.com/videojs/video.js/pull/1312))
+* Exported missing Player API methods (remainingTime, supportsFullScreen, enterFullWindow, exitFullWindow, preload) ([view](https://github.com/videojs/video.js/pull/1328))
 
 --------------------